refactor: Move config one level up into its own folder
